+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Added
+- Configurable ignored-apps list (via their bundle IDs)
 - Skip also clipboard items marked as `org.nspasteboard.TransientType` (e.g. OTP codes or passwords used by password manager's browser extensions)
 
 ## [0.8.0] - 2026-03-26